Address https://github.com/digital-asset/daml/pull/6298#discussion_r438779291
